fix：解决主页面titleBar的一个显示问题&feat：对pages文件夹进行规整
