+++ src/pages/routes/dataStatistics.vue
